max colour cast change

+++ src/include/image.hpp
@@ -139,7 +139,7 @@ class Image {
         if (data.size() == width * height) {
             os << "P6" << std::endl
                << width << ' ' << height << std::endl
-               << static_cast<uint64_t>(utils::MAX_COLOUR) << std::endl;
+               << static_cast<uint16_t>(utils::MAX_COLOUR) << std::endl;
 
             for (Colour const & c : data) {
                 c.write(os);
